--- test_autofit/unit/database/query/test_bad_synax.py
@@ -1,15 +1,0 @@
-import pytest
-
-
-def test_already_compared(
-        aggregator
-):
-    with pytest.raises(
-            AssertionError
-    ):
-        print((aggregator.centre == 1) == 1)
-
-    with pytest.raises(
-            AssertionError
-    ):
-        print((aggregator.centre == 1).intesity)
